+++ pkg/api/plugin_resource_test.go
@@ -26,7 +26,6 @@ import (
 	"github.com/grafana/grafana/pkg/plugins/manager/registry"
 	"github.com/grafana/grafana/pkg/plugins/manager/signature"
 	"github.com/grafana/grafana/pkg/plugins/manager/sources"
-	"github.com/grafana/grafana/pkg/plugins/manager/store"
 	"github.com/grafana/grafana/pkg/plugins/pluginscdn"
 	"github.com/grafana/grafana/pkg/services/accesscontrol"
 	datasources "github.com/grafana/grafana/pkg/services/datasources/fakes"
@@ -35,6 +34,7 @@ import (
 	pluginClient "github.com/grafana/grafana/pkg/services/plugins/client"
 	"github.com/grafana/grafana/pkg/services/plugins/plugincontext"
 	pluginSettings "github.com/grafana/grafana/pkg/services/plugins/pluginsettings/service"
+	"github.com/grafana/grafana/pkg/services/plugins/store"
 	"github.com/grafana/grafana/pkg/services/pluginsintegration"
 	"github.com/grafana/grafana/pkg/services/quota/quotatest"
 	"github.com/grafana/grafana/pkg/services/user"
