+++ pkg/services/plugins/store/store_test.go
